refactor: be - AppService 생성 및 API 요청 로직 분리 #66
