+++ backend/src/util/socket/object.dto.ts
